Fix Auth
